[#20250701] Separate CLI and Formatters
