+++ lib/tryouts/cli.rb
@@ -1,6 +1,10 @@
 # lib/tryouts/cli.rb
 
 require 'optparse'
+
+require_relative 'cli/opts'
+require_relative 'cli/formatters'
+
 require_relative 'prism_parser'
 require_relative 'testbatch'
 require_relative 'translators/rspec_translator'
@@ -28,26 +32,6 @@ class Tryouts
         generate_only: false,
       },
     }.freeze
-
-    HELP = <<~HELP
-
-      Framework Defaults:
-        Tryouts:    Shared context (state persists across tests)
-        RSpec:      Fresh context (each test isolated)
-        Minitest:   Fresh context (each test isolated)
-
-      Examples:
-        try test_try.rb                          # Tryouts test runner with shared context
-        try --rspec test_try.rb                  # RSpec with fresh context
-        try --direct --shared-context test_try.rb # Explicit shared context
-        try --generate-rspec test_try.rb         # Output RSpec code only
-        try --inspect test_try.rb                # Inspect file structure and validation
-
-      File Format:
-        ## Test description       # Test case marker
-        code_to_test             # Ruby code
-        #=> expected_result       # Expectation
-    HELP
 
     def initialize
       Tryouts.debug 'CLI#initialize: Initializing CLI.'
@@ -247,63 +231,6 @@ class Tryouts
       Tryouts.debug 'CLI#show_grand_total: Grand total displayed.'
     end
 
-    class << self
-      def parse_args(args)
-        Tryouts.debug "CLI::parse_args: Parsing arguments: #{args.inspect}"
-        options = {}
 
-        parser = OptionParser.new do |opts|
-          opts.banner = "Usage: try [OPTIONS] FILE...\n\nModern Tryouts test runner with framework translation"
-
-          opts.separator "\nFramework Options:"
-          opts.on('--direct', 'Direct execution with TestBatch (default)') { options[:framework] = :direct }
-          opts.on('--rspec', 'Use RSpec framework') { options[:framework]                        = :rspec }
-          opts.on('--minitest', 'Use Minitest framework') { options[:framework]                  = :minitest }
-
-          opts.separator "\nGeneration Options:"
-          opts.on('--generate-rspec', 'Generate RSpec code only') do
-            options[:framework]     = :rspec
-            options[:generate_only] = true
-          end
-          opts.on('--generate-minitest', 'Generate Minitest code only') do
-            options[:framework]     = :minitest
-            options[:generate_only] = true
-          end
-          opts.on('--generate', 'Generate code only (use with --rspec/--minitest)') do
-            options[:generate_only] = true
-            options[:framework]   ||= :rspec
-          end
-
-          opts.separator "\nExecution Options:"
-          opts.on('--shared-context', 'Override default context mode') { options[:shared_context]       = true }
-          opts.on('--no-shared-context', 'Override default context mode') { options[:shared_context]    = false }
-          opts.on('-v', '--verbose', 'Show detailed test output with line numbers') { options[:verbose] = true }
-          opts.on('-f', '--fails', 'Show only failing tests (with --verbose)') { options[:fails_only]   = true }
-
-          opts.separator "\nInspection Options:"
-          opts.on('-i', '--inspect', 'Inspect file structure without running tests') { options[:inspect] = true }
-
-          opts.separator "\nGeneral Options:"
-          opts.on('-V', '--version', 'Show version') { options[:version] = true }
-          opts.on('-D', '--debug', 'Enable debug mode') { Tryouts.debug  = true }
-          opts.on('-h', '--help', 'Show this help') do
-            Tryouts.debug 'CLI::parse_args: Help flag detected. Showing help and exiting.'
-            puts opts
-            exit 0
-          end
-
-          opts.separator HELP.freeze
-        end
-
-        files = parser.parse(args)
-        Tryouts.debug "CLI::parse_args: Parsed files: #{files.inspect}, options: #{options.inspect}"
-        [files, options]
-      rescue OptionParser::InvalidOption => ex
-        Tryouts.debug "CLI::parse_args: Invalid option error: #{ex.message}"
-        warn "Error: #{ex.message}"
-        warn "Try 'try --help' for more information."
-        exit 1
-      end
-    end
   end
 end
