+++ lib/tryouts/cli/formatters.rb
@@ -1,0 +1,6 @@
+# lib/tryouts/cli/formatters.rb
+
+require_relative 'formatters/base'
+require_relative 'formatters/compact'
+require_relative 'formatters/quiet'
+require_relative 'formatters/verbose'
